Move compilation tests of all toolchains up to prevent too long shebang
When installing to a deeply nested dir hierarchy, shebangs in shell
scripts will become too long and fail to run (hard limit of 127).

Simply move tests up in modules to prevent this.
